Update tests and ForkedArray cancel
